+++ src/components/list/list.css
@@ -1,0 +1,18 @@
+.list-container {
+    max-width: 280px;
+    margin: 0 auto;
+    padding: 10px;
+    border: solid black;
+}
+
+.list-container ul {
+    list-style-type: none;
+    padding: 0;
+}
+
+.list-container li {
+    margin: 5px 0;
+    word-wrap: break-word;
+    border: solid black;
+    height: 100px;
+}
